# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cuanku API Documentation",
        "description": "API documentation for Cuanku Financial Management App",
        "contact": {
            "email": "admin@cuanku.web.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/budgets": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "List all budgets",
                "operationId": "a661446b37868c87cb5dc8d0022e158e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BudgetResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Create a new budget",
                "operationId": "52ec07c758938927f1afad3070db7c54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Budget created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/budgets/{budget}": {
            "put": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Update an existing budget",
                "operationId": "3263828800c1f0df054f144b707da6fb",
                "parameters": [
                    {
                        "name": "budget",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Budget updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Delete a budget",
                "operationId": "5fbb4ad53a48ed0152e5cd1554da5f3d",
                "parameters": [
                    {
                        "name": "budget",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budget deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard data",
                "operationId": "a92dff3ea055c0a594f425e852840a68",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sum": {
                                            "type": "object"
                                        },
                                        "goals": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GoalResource"
                                            }
                                        },
                                        "incomes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IncomeResource"
                                            }
                                        },
                                        "expenses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ExpenseResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationError": {
                "title": "Validation Error",
                "description": "Validation error response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "The field_name is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Unauthorized": {
                "title": "Unauthorized",
                "description": "Unauthorized error response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "title": "Not Found",
                "description": "Resource not found response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model."
                    }
                },
                "type": "object"
            },
...
            "Success": {
                "title": "Success",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully."
                    }
                },
                "type": "object"
            },
...
            "BudgetRequest": {
                "title": "Budget Request",
                "description": "Budget request body data",
                "required": [
                    "detail",
                    "nominal",
                    "month",
                    "year",
                    "type"
                ],
                "properties": {
                    "detail": {
                        "type": "string"
                    },
                    "nominal": {
                        "type": "number"
                    },
                    "month": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "BudgetResource": {
                "title": "Budget Resource",
                "description": "Budget resource representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "nominal": {
                        "type": "number"
                    },
                    "month": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ExpenseResource": {
                "title": "Expense Resource",
                "description": "Expense resource representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "description": {
                        "type": "string"
                    },
                    "nominal": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "month": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "payment": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "typeDetail": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "detail": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "GoalResource": {
                "title": "Goal Resource",
                "description": "Goal resource representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "percentage": {
                        "type": "number",
                        "format": "float"
                    },
                    "nominal": {
                        "type": "number"
                    },
                    "monthly_saving": {
                        "type": "number"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date"
                    },
                    "beginning_balance": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "IncomeResource": {
                "title": "Income Resource",
                "description": "Income resource representation",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "nominal": {
                        "type": "number"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "month": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "budget": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "detail": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}